requirements.txt
fastapi[standard]
google-genai
httpx
python-dotenv